+++ packages/carp_context_package/lib/context.dart
@@ -9,7 +9,7 @@ library context;
 import 'dart:async';
 import 'dart:math' as math;
 import 'package:json_annotation/json_annotation.dart';
-import 'package:activity_recognition_flutter/activity_recognition_flutter.dart';
+// import 'package:activity_recognition_flutter/activity_recognition_flutter.dart';
 import 'package:weather/weather.dart';
 import 'package:openmhealth_schemas/openmhealth_schemas.dart' as omh;
 import 'package:permission_handler/permission_handler.dart';
@@ -19,6 +19,9 @@ import 'package:mobility_features/mobility_features.dart';
 // import 'package:geolocator/geolocator.dart';
 import 'package:location/location.dart' as location;
 
+import 'package:flutter_activity_recognition/flutter_activity_recognition.dart';
+//     as ar;
+
 import 'package:carp_core/carp_core.dart';
 import 'package:carp_mobile_sensing/carp_mobile_sensing.dart';
 
@@ -26,7 +29,7 @@ part 'src/location_manager.dart';
 part 'src/activity/activity_datum.dart';
 part 'src/activity/activity_probe.dart';
 part 'src/location/location_datum.dart';
-part 'src/location/location_probe.dart';
+part 'src/location/location_probes.dart';
 part 'src/location/location_measure.dart';
 part 'src/weather/weather_datum.dart';
 part 'src/weather/weather_measure.dart';