+++ packages/carp_context_package/lib/src/activity/activity_datum.dart
@@ -10,6 +10,21 @@ part of context;
 /// Holds activity information.
 @JsonSerializable(fieldRename: FieldRename.snake, includeIfNull: false)
 class ActivityDatum extends Datum {
+  // static Map<ar.ActivityType, ActivityType> _activityTypeMap = {
+  //   ar.ActivityType.IN_VEHICLE: ActivityType.IN_VEHICLE,
+  //   ar.ActivityType.ON_BICYCLE: ActivityType.ON_BICYCLE,
+  //   ar.ActivityType.RUNNING: ActivityType.RUNNING,
+  //   ar.ActivityType.STILL: ActivityType.STILL,
+  //   ar.ActivityType.UNKNOWN: ActivityType.UNKNOWN,
+  //   ar.ActivityType.WALKING: ActivityType.WALKING,
+  // };
+
+  static Map<ActivityConfidence, int> _confidenceLevelMap = {
+    ActivityConfidence.HIGH: 100,
+    ActivityConfidence.MEDIUM: 70,
+    ActivityConfidence.LOW: 40,
+  };
+
   DataFormat get format =>
       DataFormat.fromString(ContextSamplingPackage.ACTIVITY);
 
@@ -20,10 +35,15 @@ class ActivityDatum extends Datum {
         type = map['type'],
         super();
 
-  factory ActivityDatum.fromActivity(ActivityEvent activityEvent) =>
-      ActivityDatum(
-        activityEvent.type,
-        activityEvent.confidence,
+  // factory ActivityDatum.fromActivityEvent(ActivityEvent activityEvent) =>
+  //     ActivityDatum(
+  //       activityEvent.type,
+  //       activityEvent.confidence,
+  //     );
+
+  factory ActivityDatum.fromActivity(Activity activity) => ActivityDatum(
+        activity.type,
+        _confidenceLevelMap[activity.confidence] ?? 0,
       );
 
   factory ActivityDatum.fromJson(Map<String, dynamic> json) =>
@@ -43,8 +63,6 @@ class ActivityDatum extends Datum {
   /// * WALKING - The device is on a user who is walking.
   /// * RUNNING - The device is on a user who is running.
   /// * STILL - The device is still (not moving).
-  /// * TILTING - The device angle relative to gravity changed significantly.
-  /// * UNKNOWN - Unable to detect the current activity.
   ///
   /// The types above are adopted from the Android activity recognition API.
   /// On iOS the following mapping takes place:
@@ -54,7 +72,11 @@ class ActivityDatum extends Datum {
   /// * running => RUNNING
   /// * automotive => IN_VEHICLE
   /// * cycling => ON_BICYCLE
-  /// * unknown => UNKNOWN
+  ///
+  /// Note that the [ActivityProbe] discard some AR events, which include:
+  ///  * [ActivityType.UNKNOWN]
+  ///  * [ActivityType.TILTING]
+  ///  * Activities with a low confidence level (<50%)
   ActivityType type;
 
   /// Activity [type] as a string.