+++ packages/carp_context_package/lib/src/mobility/mobility_probe.dart
@@ -1,6 +1,6 @@
 part of context;
 
-/// Collects mobility features using the [MobilityFactory] API.
+/// Collects mobility features using the [MobilityFeatures] API.
 class MobilityProbe extends StreamProbe {
   @override
   void onInitialize(Measure measure) {
@@ -18,13 +18,14 @@ class MobilityProbe extends StreamProbe {
     await LocationManager().configure(measure as LocationConfiguration);
 
     // start the location data stream from the LocationManager
-    Stream<LocationSample> stream = LocationManager().onLocationChanged.map(
-        (loc) => LocationSample(
+    Stream<LocationSample> locationStream = LocationManager()
+        .onLocationChanged
+        .map((loc) => LocationSample(
             GeoLocation(loc.latitude!, loc.longitude!), DateTime.now()));
 
     // feed the location data stream to the MobilityFeatures singleton
     // which in turn produce [MobilityContext]s
-    await MobilityFeatures().startListening(stream);
+    await MobilityFeatures().startListening(locationStream);
 
     super.onResume();
   }