+++ packages/carp_context_package/test/carp_context_package_test.dart
@@ -5,7 +5,8 @@ import 'package:test/test.dart';
 import 'package:carp_core/carp_core.dart';
 import 'package:carp_mobile_sensing/carp_mobile_sensing.dart';
 import 'package:carp_context_package/context.dart';
-import 'package:activity_recognition_flutter/activity_recognition_flutter.dart';
+// import 'package:activity_recognition_flutter/activity_recognition_flutter.dart';
+import 'package:flutter_activity_recognition/flutter_activity_recognition.dart';
 
 String _encode(Object object) =>
     const JsonEncoder.withIndent(' ').convert(object);